moving state converters into solver_classes.py
