+++ qiskit_dynamics/solve.py
@@ -281,53 +281,3 @@ def solve_lmde(
             results.y = results.y.T
 
     return results
-
-
-def initial_state_converter(
-    obj: Any, return_class: bool = False
-) -> Union[Array, Tuple[Array, Type]]:
-    """Convert initial state object to an Array.
-
-    Args:
-        obj: An initial state.
-        return_class: Optional. If True return the class to use
-                      for converting the output y Array.
-
-    Returns:
-        Array: the converted initial state if ``return_class=False``.
-        tuple: (Array, class) if ``return_class=True``.
-    """
-    # pylint: disable=invalid-name
-    y0_cls = None
-    if isinstance(obj, Array):
-        y0, y0_cls = obj, None
-    if isinstance(obj, QuantumState):
-        y0, y0_cls = Array(obj.data), obj.__class__
-    elif isinstance(obj, QuantumChannel):
-        y0, y0_cls = Array(SuperOp(obj).data), SuperOp
-    elif isinstance(obj, (BaseOperator, Gate, QuantumCircuit)):
-        y0, y0_cls = Array(Operator(obj.data)), Operator
-    else:
-        y0, y0_cls = Array(obj), None
-    if return_class:
-        return y0, y0_cls
-    return y0
-
-
-def final_state_converter(obj: Any, cls: Optional[Type] = None) -> Any:
-    """Convert final state Array to custom class.
-
-    Args:
-        obj: final state Array.
-        cls: Optional. The class to convert to.
-
-    Returns:
-        Any: the final state.
-    """
-    if cls is None:
-        return obj
-
-    if issubclass(cls, (BaseOperator, QuantumState)) and isinstance(obj, Array):
-        return cls(obj.data)
-
-    return cls(obj)
